ui/Point: allow constructing with unsigned integer

+++ src/FileListPage.cxx
@@ -540,7 +540,7 @@ FileListPage::PaintStatusBarOverride(const Window window) const noexcept
 	const unsigned duration_width = duration_string.size();
 
 	SelectStyle(window, Style::STATUS_TIME);
-	window.String({(int)window.GetWidth() - (int)duration_width, 0}, duration_string);
+	window.String({window.GetWidth() - duration_width, 0u}, duration_string);
 
 	window.RefreshNoOut();
